fix: process_withdrawal and add tests
