+++ testing/ef-tests/src/macros/mod.rs
@@ -1,0 +1,2 @@
+pub mod operations;
+pub mod ssz_static;
